Move BaseExportResource to trips, since it's only used there now
